requirements.txt
pandas==2.2.3
psycopg==3.2.6
python-dotenv==1.1.0
python_Levenshtein==0.27.1
Requests==2.32.3
selenium==4.31.0
validators==0.22.0
